requirements.txt
pytest
pytest-asyncio
playwright
python-dotenv